Adding new folder structure for flows
